requirements.txt
paho-mqtt==1.6.1
streamlit==1.22.0
python-dotenv==1.0.0